requirements.txt
certifi==2023.11.17
charset-normalizer==3.2.0
idna==3.6
numpy==1.26.2
opencv-python==4.9.0.80
pyTelegramBotAPI==4.12.0
python-dotenv==1.0.0
requests==2.31.0
telebot==0.0.5
urllib3==2.0.4
